Original Upload
"Create a Mad Libs program that reads in text files and lets the user add their own text anywhere the word ADJECTIVE, NOUN, ADVERB, or VERB appears in the text file. The program would find these occurrences and prompt the user to replace them. The results should be printed to the screen and saved to a new text file."